+++ src/IdentifierNode.hpp
@@ -6,7 +6,7 @@ class IdentifierNode : public Node {
 public:
     std::string name;
     std::string index_name;
-    int index_value;
+    ll index_value;
     bool is_array = false;
     bool is_iterator = false;
     bool is_pointer = false; // It can be recognized later, not in grammar
